remplace int by Py_ssize_t and for rank8

+++ skimage/rank/_core8.pxd
@@ -4,7 +4,7 @@ cimport numpy as np
 # 8 bit core kernel
 #---------------------------------------------------------------------------
 
-cdef inline _core8(np.uint8_t kernel(int*, float, np.uint8_t),
+cdef inline _core8(np.uint8_t kernel(Py_ssize_t*, float, np.uint8_t),
 np.ndarray[np.uint8_t, ndim=2] image,
 np.ndarray[np.uint8_t, ndim=2] selem,
 np.ndarray[np.uint8_t, ndim=2] mask,
